Cleanining up the pull request #1 by avoiding code repetition.
I'm just tentatively committing pull request #1 as it's been laying dormant for a while now, and the use case seemed legitimate to me.

I do think the proper implementation is to introduce an extension point that defines various criteria, and I have a mixed feeling about having 9 options in the UI like this, but on the positive side this has limited backward compatibility overhead in the future. To whoever maintains this plugin, please feel free to adjust this as you see fit.